added home component

+++ src/app/register/register.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { UserService } from '../user.service';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-register',
@@ -7,45 +7,55 @@ import { UserService } from '../user.service';
   styleUrls: ['./register.component.scss']
 })
 export class RegisterComponent implements OnInit {
- user:any={}
-  constructor(private userService: UserService) {
+  name:String='';
+  surname:String='';
+  password:String='';
+  confirmPassword:String='';
+  gender:String='';
+  email:String='';
+  cell:String='';
+  question:String='';
+  answer:String='';
+
+ questions: string[] = ['What is your ID number?', 'what year did you finish school?'];
+ registerUser: any[] = [];
+  constructor(private router: Router) {
   
-    
   }
 
   ngOnInit(): void {
-    console.log("object from service",this.userService)
-    this.user= this.userService
-    let storedUsers = localStorage.getItem('registerUser');
-    if (storedUsers) {
-      this.registerUser = JSON.parse(storedUsers);
-    } else {
-      this.registerUser = [];
-    }
+   
+   
   }
-
-
-  questions: string[] = ['What is your ID number?', 'what year did you finish school?'];
-  registerUser: any[] = [];
 
   register() {
     for(let i = 0; i < this.registerUser.length; i++){
-      if (this.registerUser[i].email === ""){
-        alert("This User already exists");
-      }
+      if (this.registerUser[i].email === this.email){
+        console.log("This User already exists");
+      } 
     }
-    //if (!this.name || !this.surname || !this.password || !this.confirmPassword || !this.gender || !this.email || !this.cell || !this.question || !this.answer) {
-      //alert("Please fill in ALL fields!");
-     {
-      
-      this.registerUser.push(this.userService);
+    //console.log(this.name, this.surname, this.password, this.confirmPassword, this.gender, this.email, this.cell, this.question, this.answer);
+    if (this.name === '' || this.surname === '' || this.password === '' || this.confirmPassword === '' || this.gender === '' || this.email === '' || this.cell === '' || this.question === '' || this.answer === '') {
+      alert('Please fill in ALL fields!');
+    }else{
+      let user ={
+        name: this.name,
+        surname: this.surname,
+        email: this.email,
+        password: this.password,
+        confirmPassword: this.confirmPassword,
+        cell: this.cell,
+        question: this.question,
+        answer: this.answer,
+        gender: this.gender,
+      };
+      this.registerUser.push(user);
       localStorage.setItem('registerUser', JSON.stringify(this.registerUser));
       alert("Registration successful!");
-      this.directToLogin();
-    }
-  }
-
-  directToLogin(){
-    document.location.href = "http://localhost:4200/login"
-  }
+      this.router.navigate(['/login']);
+      
+  
+  
+  } 
+}
 }
